Extract CustomCTAStep
